Рабочая версия

+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImp.java
@@ -1,15 +1,23 @@
 package ru.kata.spring.boot_security.demo.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import ru.kata.spring.boot_security.demo.model.User;
 import ru.kata.spring.boot_security.demo.repository.UserRepository;
-import java.util.List;
+
+import java.util.HashSet;
 import java.util.Set;
 
 @Service
-public class UserServiceImp implements UserService {
-
+public class UserServiceImp implements UserService, UserDetailsService {
+    @Autowired
+    BCryptPasswordEncoder bCryptPasswordEncoder;
+    @Autowired
+    RoleService roleService;
     private final UserRepository userRepository;
 
     @Autowired
@@ -18,15 +26,27 @@ public class UserServiceImp implements UserService {
     }
 
     @Override
-    public void saveOrUpdateUser(User user) {
+    public boolean saveOrUpdateUser(User user) {
+        if (user.getId() == null && userRepository.findByUsername(user.getUsername()) != null) {
+            return false;
+        }
+        user.getRolesSet().add(roleService.getAllRoles().stream().filter(r -> r.getNameRole()
+                .equals("ROLE_USER")).findFirst().orElse(null));
+        if (user.getId() == null) {
+            user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+        }
         userRepository.save(user);
+        return true;
     }
 
     @Override
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public boolean deleteUser(Long id) {
+        if (userRepository.findById(id).isPresent()) {
+            userRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
-
 
     @Override
     public User getUserById(Long id) {
@@ -35,7 +55,24 @@ public class UserServiceImp implements UserService {
 
     @Override
     public Set<User> getAllUsers() {
-        return (Set<User>) userRepository.findAll();
+        Set<User> set = new HashSet<>();
+        Iterable<User> iterable = userRepository.findAll();
+        iterable.forEach(set::add);
+        return set;
     }
 
+    @Override
+    public User findByUsername(String login) {
+        return userRepository.findByUsername(login);
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        User user = userRepository.findByUsername(username);
+        if (user == null) {
+            throw new UsernameNotFoundException("User not found");
+        }
+        return user;
+    }
 }
+
